requirements.txt
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1